+++ src/com/ihgoo/allinone/util/IOUtil.java
@@ -3,6 +3,7 @@ package com.ihgoo.allinone.util;
 import java.io.BufferedInputStream;
 import java.io.BufferedOutputStream;
 import java.io.BufferedReader;
+import java.io.Closeable;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
@@ -12,11 +13,34 @@ import java.io.InputStream;
 import java.io.InputStreamReader;
 
 import android.content.Context;
+import android.database.Cursor;
 
 /**
  * Created by ihgoo on 2015/6/16.
  */
 public class IOUtil {
+	
+	
+	   private IOUtil() {
+	    }
+
+	    public static void closeQuietly(Closeable closeable) {
+	        if (closeable != null) {
+	            try {
+	                closeable.close();
+	            } catch (Throwable e) {
+	            }
+	        }
+	    }
+
+	    public static void closeQuietly(Cursor cursor) {
+	        if (cursor != null) {
+	            try {
+	                cursor.close();
+	            } catch (Throwable e) {
+	            }
+	        }
+	    }
 
     /**
      * @return 导出数据流
