--- src/com/ihgoo/allinone/util/IOUtils.java
@@ -1,29 +1,0 @@
-package com.ihgoo.allinone.util;
-
-import android.database.Cursor;
-
-import java.io.Closeable;
-
-public class IOUtils {
-
-    private IOUtils() {
-    }
-
-    public static void closeQuietly(Closeable closeable) {
-        if (closeable != null) {
-            try {
-                closeable.close();
-            } catch (Throwable e) {
-            }
-        }
-    }
-
-    public static void closeQuietly(Cursor cursor) {
-        if (cursor != null) {
-            try {
-                cursor.close();
-            } catch (Throwable e) {
-            }
-        }
-    }
-}
